Update test_server to send in real time.
Or at least attempt to! This measures the real time taken to send
the messages between two timestamps and uses the ratio between
this real time and the data time to update the time interval
between subquent packet in the future. Often NMEA data can be
bursty, so in reality this only gives an approximate correlation
between the passage of real time and data time.