Refactor displayable to a cell class heirarchy.
Previously the state module contained a "displayable" heirarchy
that could provide heading/unit/value strings to populate a cell.

This was a (mild) violation of separation of concerns since
these are features of the UI rather than the state and it also
constrained our ability to add other cell types.

After this commit displayable is gone and replaced by an
inheritance heirarchy of Cell classes in UI, most based on a
CellSpec. This ensures the state knows nothing about the UI,
is a more flexible framework for cells that depend on other
inputs such as history, and supports cells that need to
display something other than a pure text string.

The only user visible change should be that strings such as the
"hold to configure" prompt are centered on the whole cell rather
than the place the value would lie.