Support heading graphs.
This ended up being a massive refactor because a heading needs to
carry both a bearing and a variation in order to be formatted as
either. All the history concepts like storage, serialization, and
accumulation needed to become generic over the Value type and I
split the property and source attributes into a separate BoundValue
type to make the overwhelming number of cases where they are not
needed simpler.

As a side benefit I also wrote a lot more tests.